refactor #149: use new mock for api data
